Rename vtltape.pem to vtltape_pem.h
It is not a PEM-format file, it is a C header containing a
representation of a PEM-formatted chunk of data.
